Split MapTypes.kt into one file per type in map.model package
