+++ composeApp/src/androidInstrumentedTest/kotlin/com/jordankurtz/piawaremobile/SettingsScreenAndroidTest.kt
@@ -8,6 +8,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performScrollToNode
+import com.jordankurtz.piawaremobile.map.cache.TileCache
 import com.jordankurtz.piawaremobile.model.Async
 import com.jordankurtz.piawaremobile.settings.Settings
 import com.jordankurtz.piawaremobile.settings.SettingsViewModel
@@ -33,6 +34,7 @@ class SettingsScreenAndroidTest {
 
         return SettingsViewModel(
             settingsService = settingsService,
+            tileCache = mock<TileCache>(),
         )
     }
