+++ composeApp/src/commonMain/kotlin/com/jordankurtz/piawaremobile/map/model/LatLon.kt
@@ -1,0 +1,3 @@
+package com.jordankurtz.piawaremobile.map.model
+
+data class LatLon(val latitude: Double, val longitude: Double)
