+++ composeApp/src/commonMain/kotlin/com/jordankurtz/piawaremobile/map/model/MapBounds.kt
@@ -1,0 +1,8 @@
+package com.jordankurtz.piawaremobile.map.model
+
+data class MapBounds(
+    val north: Double,
+    val south: Double,
+    val east: Double,
+    val west: Double,
+)
