+++ composeApp/src/commonMain/kotlin/com/jordankurtz/piawaremobile/map/model/MapPosition.kt
@@ -1,0 +1,3 @@
+package com.jordankurtz.piawaremobile.map.model
+
+data class MapPosition(val latitude: Double, val longitude: Double, val zoom: Double)
